Fixpoint storage size

+++ Sierp/main.cpp
@@ -19,7 +19,7 @@ int render(grid<uint8_t> &output){
     
     grid<uint8_t> *current = &a;
     grid<uint8_t> *next = &b;
-    sierp::Point<uint8_t> points[3];
+    sierp::Point<size_t> points[3];
     points[0].x = 0; points[0].y = HEIGHT;
     points[1].x = WIDTH; points[1].y = HEIGHT;
     points[2].x = WIDTH/2; points[2].y = 0;
